chapters/2020/assets/localisation/codes/Rimel-i18n-FMPP-master/requirements.txt
attrs==19.3.0
Automat==0.8.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
cycler==0.10.0
get==2019.4.13
gitdb2==2.0.6
GitPython==3.0.5
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
kiwisolver==1.1.0
lizard==1.17.3
lxml==4.4.2
matplotlib==3.1.3
numpy==1.18.1
parsel==1.5.2
post==2019.4.13
Protego==0.1.16
public==2019.4.13
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyDispatcher==2.0.5
PyDriller==1.11
PyHamcrest==2.0.0
pyOpenSSL==19.1.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
query-string==2019.4.13
queuelib==1.5.0
requests==2.22.0
Scrapy==1.8.0
service-identity==18.1.0
six==1.13.0
smmap==3.0.1
smmap2==2.0.5
Twisted==19.10.0
urllib3==1.25.8
w3lib==1.21.0
zope.interface==4.7.1
